feat: testing javascript lint
