+++ libs/frontend/view/components/editors/CodeMirrorEditor.tsx
@@ -1,18 +1,34 @@
+import 'eslint-linter-browserify'
 import { CodeMirrorLanguage } from '@codelab/shared/abstract/codegen'
 import { css } from '@codemirror/lang-css'
-import { javascript } from '@codemirror/lang-javascript'
+import { esLint, javascript } from '@codemirror/lang-javascript'
 import { json } from '@codemirror/lang-json'
+import { StreamLanguage } from '@codemirror/language'
+import { linter, lintGutter } from '@codemirror/lint'
+import { Extension } from '@uiw/react-codemirror'
+import { graphql } from 'codemirror-graphql/cm6-legacy/mode'
 import React from 'react'
-import { CodeMirrorInput } from './codeMirrorInput'
-import { CodeMirrorEditorProps } from './codeMirrorInput/types'
+import { CodeMirrorInput, CodeMirrorInputProps } from './codeMirrorInput'
 
-const languageExtension = {
+export interface CodeMirrorEditorProps extends CodeMirrorInputProps {
+  language?: CodeMirrorLanguage
+}
+
+const languageExtension: { [lang in CodeMirrorLanguage]: Array<Extension> } = {
   [CodeMirrorLanguage.Css]: [css()],
-  [CodeMirrorLanguage.Javascript]: [javascript()],
-  [CodeMirrorLanguage.Typescript]: [javascript({ typescript: true })],
+  [CodeMirrorLanguage.Javascript]: [
+    linter(esLint(new eslint.Linter())),
+    lintGutter(),
+    javascript(),
+  ],
+  [CodeMirrorLanguage.Typescript]: [
+    linter(esLint(new eslint.Linter())),
+    lintGutter(),
+    javascript({ typescript: true }),
+  ],
   [CodeMirrorLanguage.Json]: [json()],
   [CodeMirrorLanguage.CssInJs]: [css()],
-  [CodeMirrorLanguage.Graphql]: [json()],
+  [CodeMirrorLanguage.Graphql]: [StreamLanguage.define(graphql)],
 }
 
 export const CodeMirrorEditor = (props: CodeMirrorEditorProps) => {
@@ -23,7 +39,7 @@ export const CodeMirrorEditor = (props: CodeMirrorEditorProps) => {
       // eslint-disable-next-line react/jsx-props-no-spreading
       {...props}
       extensions={
-        language ? [...languageExtension[language], ...extensions] : extensions
+        language ? languageExtension[language].concat(extensions) : extensions
       }
       shouldDisableNewLines={false}
     />
