+++ package.json
@@ -55,6 +55,7 @@
     "@codemirror/lang-html": "^0.20.0",
     "@codemirror/lang-javascript": "^0.20.0",
     "@codemirror/lang-json": "^0.20.0",
+    "@codemirror/legacy-modes": "^6.1.0",
     "@dnd-kit/core": "^6.0.5",
     "@emotion/react": "11.10.0",
     "@emotion/server": "11.10.0",
@@ -101,7 +102,7 @@
     "class-transformer": "0.5.1",
     "class-validator": "0.13.2",
     "codemirror": "5",
-    "codemirror-graphql": "^1.3.0",
+    "codemirror-graphql": "^1.3.2",
     "core-js": "3.20.2",
     "cors": "2.8.5",
     "cross-fetch": "3.1.4",
@@ -112,6 +113,7 @@
     "deepmerge": "4.2.2",
     "env-var": "7.1.1",
     "enzyme": "3.11.0",
+    "eslint-linter-browserify": "^8.22.0",
     "fp-ts": "2.11.7",
     "framer-motion": "6.3.3",
     "fuse.js": "6.5.3",
